Switch to create-react-app, its files were added
and the previous src folder was deleted locally to be able to run the
create-react-app scripts and test the functionality but this is not commited
with this commit here. As many of those previous src files will be modified to
be compatible with create-react-app and only some of them will be deleted.

The rationale for this switch is that c-r-a has a laid out path for suitable
and thus optimized deployment especially regarding bundle size.